Move drawChart to its own JS file
